# vail_scraper/routers/api/v2/openapi.yaml
openapi: "3.1.0"
info:
  title: "Vail scraper API"
  version: "2"
  summary: "Automatically scrapes vail's API and returns it in a usable format"
  description: |
    Rate limits are per-ip, however overrides may be added if you ask.
  license:
    name: "GPL v3"
    identifier: "GPL-3.0"
  contact:
    name: "Discord"
    url: "https://discord.com/users/397745647723216898"
servers:
  - description: "Production"
    url: "https://vail-scraper.farfrom.world/api/v2"
  - description: "Local instance"
    url: "http://localhost:8000"
...
paths:
...
  /game/user-count/timeseries:
    get:
      operationId: "getGameUserCountTimeseries"
      summary: "Get how many users vail has over time"
      description: |
        # Info
        Get how many people have a account on VAIL
        
        # What counts as a VAIL user?
        This only counts people who have started vail once.
        
        Users who use family sharing will get one account counted per "account" in the family share
        
        Users who create "full" accounts through https://login.vailvr.com will only count once despite being on steam and oculus.
        
        # Sorting of response items
        Providing `after`: lowest matching timestamp -> highest
        
        Providing `before`: highest matching timestamp -> lowest
        
        Providing nothing: most recent scrape -> oldest scrape
        
        # Rate limits
        This endpoint is currently not rate limited.
      tags: ["game"]
      
      parameters:
        - name: "before"
          description: |
            This is mutually exclusive with `after`
            
            **NOTE** Please read the request description regarding sorting
          in: "query"
          schema:
            $ref: "#/components/schemas/Timestamp"
            description: "if not sent this will default to the current timestamp"
        - name: "after"
          in: "query"
          description: |
            This is mutually exclusive with `before`
            
            **NOTE** Please read the request description regarding sorting
          schema:
            $ref: "#/components/schemas/Timestamp"
            default: 0
        - name: "limit"
          in: "query"
          schema:
            type: "number"
            default: 100
            minimum: 0
            maximum: 100
      responses:
        "200":
          description: "Success"
          content:
             application/json:
               schema:
                 type: "object"
                 required:
                   - "items"
                 properties:
                   items:
                     type: "array"
                     items:
                       $ref: "#/components/schemas/UserCountTimeseriesEntry"
        "400":
           description: "You did something stupid"
           content:
             application/json:
               schema:
                 oneOf:
                   - $ref: "#/components/schemas/QueryParamInvalidError"
components:
...
  schemas:
...
    Timestamp:
      type: "number"
      format: "unix timestamp"
      example: 1717260094.5
    StringTimestamp:
      type: "string"
      format: "unix timestamp"
      example: "1717260094.5"
    
    # User stats
...
    UserCountTimeseriesEntry:
      type: "object"
      required:
        - "timestamp"
        - "count"
      properties:
        timestamp:
          $ref: "#/components/schemas/StringTimestamp"
        count:
          type: "integer"
    
          
    
    # Errors
...
    QueryParamInvalidError:
      type: "object"
      required:
        - "code"
        - "detail"
        - "field"
      properties:
        code:
          type: "string"
          enum:
            - "query_parameter_invalid"
        detail:
          type: "string"
          example: "failed to parse the after parameter"
        field:
          type: "string"